pipeline/requirements.txt
beautifulsoup4
requests
sentence_transformers
pymongo
python-dotenv